Refactor chat runtime modules
